requirements/development.txt
-r production.txt
Ipython[all]
pep8
django-debug-toolbar
django-extensions
django-custom-user
